+++ tests/Integration/MembershipXmlTest.php
@@ -2,27 +2,16 @@
 
 namespace Redmine\Tests\Integration;
 
-use DOMDocument;
 use PHPUnit\Framework\TestCase;
 use Redmine\Exception\MissingParameterException;
 use Redmine\Tests\Fixtures\MockClient;
-use SimpleXMLElement;
 
 class MembershipXmlTest extends TestCase
 {
-    /**
-     * @var MockClient
-     */
-    private $client;
-
-    public function setup(): void
-    {
-        $this->client = new MockClient('http://test.local', 'asdf');
-    }
-
     public function testCreateBlank()
     {
-        $api = $this->client->getApi('membership');
+        /** @var \Redmine\Api\Membership */
+        $api = MockClient::create()->getApi('membership');
         $this->assertInstanceOf('Redmine\Api\Membership', $api);
 
         $this->expectException(MissingParameterException::class);
@@ -33,49 +22,53 @@ class MembershipXmlTest extends TestCase
 
     public function testCreateComplex()
     {
-        $api = $this->client->getApi('membership');
+        /** @var \Redmine\Api\Membership */
+        $api = MockClient::create()->getApi('membership');
         $res = $api->create('otherProject', [
             'user_id' => 1,
             'role_ids' => [1, 2],
         ]);
-        $res = json_decode($res, true);
+        $response = json_decode($res, true);
 
-        $xml = '<?xml version="1.0"?>
-<membership>
-    <user_id>1</user_id>
-    <role_ids type="array">
-        <role_id>1</role_id>
-        <role_id>2</role_id>
-    </role_ids>
-</membership>';
-        $this->assertEquals($this->formatXml($xml), $this->formatXml($res['data']));
+        $this->assertEquals('POST', $response['method']);
+        $this->assertEquals('/projects/otherProject/memberships.xml', $response['path']);
+        $this->assertXmlStringEqualsXmlString(
+            <<< XML
+            <?xml version="1.0"?>
+            <membership>
+                <user_id>1</user_id>
+                <role_ids type="array">
+                    <role_id>1</role_id>
+                    <role_id>2</role_id>
+                </role_ids>
+            </membership>
+            XML,
+            $response['data']
+        );
     }
 
     public function testUpdate()
     {
-        $api = $this->client->getApi('membership');
+        /** @var \Redmine\Api\Membership */
+        $api = MockClient::create()->getApi('membership');
         $res = $api->update(1, [
             'role_ids' => [1, 2],
         ]);
-        $res = json_decode($res, true);
+        $response = json_decode($res, true);
 
-        $xml = '<?xml version="1.0"?>
-<membership>
-    <role_ids type="array">
-        <role_id>1</role_id>
-        <role_id>2</role_id>
-    </role_ids>
-</membership>';
-        $this->assertEquals($this->formatXml($xml), $this->formatXml($res['data']));
-    }
-
-    private function formatXml($xml)
-    {
-        $dom = new DOMDocument('1.0');
-        $dom->preserveWhiteSpace = false;
-        $dom->formatOutput = true;
-        $dom->loadXML((new SimpleXMLElement($xml))->asXML());
-
-        return $dom->saveXML();
+        $this->assertEquals('PUT', $response['method']);
+        $this->assertEquals('/memberships/1.xml', $response['path']);
+        $this->assertXmlStringEqualsXmlString(
+            <<< XML
+            <?xml version="1.0"?>
+            <membership>
+                <role_ids type="array">
+                    <role_id>1</role_id>
+                    <role_id>2</role_id>
+                </role_ids>
+            </membership>
+            XML,
+            $response['data']
+        );
     }
 }