+++ tests/Integration/ProjectXmlTest.php
@@ -2,27 +2,16 @@
 
 namespace Redmine\Tests\Integration;
 
-use DOMDocument;
 use PHPUnit\Framework\TestCase;
 use Redmine\Exception\MissingParameterException;
 use Redmine\Tests\Fixtures\MockClient;
-use SimpleXMLElement;
 
 class ProjectXmlTest extends TestCase
 {
-    /**
-     * @var MockClient
-     */
-    private $client;
-
-    public function setup(): void
-    {
-        $this->client = new MockClient('http://test.local', 'asdf');
-    }
-
     public function testCreateBlank()
     {
-        $api = $this->client->getApi('project');
+        /** @var \Redmine\Api\Project */
+        $api = MockClient::create()->getApi('project');
         $this->assertInstanceOf('Redmine\Api\Project', $api);
 
         $this->expectException(MissingParameterException::class);
@@ -33,7 +22,8 @@ class ProjectXmlTest extends TestCase
 
     public function testCreateComplex()
     {
-        $api = $this->client->getApi('project');
+        /** @var \Redmine\Api\Project */
+        $api = MockClient::create()->getApi('project');
         $res = $api->create([
             'name' => 'some name',
             'identifier' => 'the_identifier',
@@ -46,28 +36,35 @@ class ProjectXmlTest extends TestCase
                 ],
             ],
         ]);
-        $res = json_decode($res, true);
+        $response = json_decode($res, true);
 
-        $xml = '<?xml version="1.0"?>
-<project>
-    <name>some name</name>
-    <identifier>the_identifier</identifier>
-    <custom_fields type="array">
-        <custom_field name="cf_name" field_format="string" id="123" multiple="true">
-            <value type="array">
-                <value>1</value>
-                <value>2</value>
-                <value>3</value>
-            </value>
-        </custom_field>
-    </custom_fields>
-</project>';
-        $this->assertEquals($this->formatXml($xml), $this->formatXml($res['data']));
+        $this->assertEquals('POST', $response['method']);
+        $this->assertEquals('/projects.xml', $response['path']);
+        $this->assertXmlStringEqualsXmlString(
+            <<< XML
+            <?xml version="1.0"?>
+            <project>
+                <name>some name</name>
+                <identifier>the_identifier</identifier>
+                <custom_fields type="array">
+                    <custom_field name="cf_name" field_format="string" id="123" multiple="true">
+                        <value type="array">
+                            <value>1</value>
+                            <value>2</value>
+                            <value>3</value>
+                        </value>
+                    </custom_field>
+                </custom_fields>
+            </project>
+            XML,
+            $response['data']
+        );
     }
 
     public function testCreateComplexWithTrackerIds()
     {
-        $api = $this->client->getApi('project');
+        /** @var \Redmine\Api\Project */
+        $api = MockClient::create()->getApi('project');
         $res = $api->create([
             'name' => 'some name',
             'identifier' => 'the_identifier',
@@ -75,44 +72,47 @@ class ProjectXmlTest extends TestCase
                 1, 2, 3,
             ],
         ]);
-        $res = json_decode($res, true);
+        $response = json_decode($res, true);
 
-        $xml = '<?xml version="1.0"?>
-<project>
-    <name>some name</name>
-    <identifier>the_identifier</identifier>
-    <tracker_ids type="array">
-        <tracker>1</tracker>
-        <tracker>2</tracker>
-        <tracker>3</tracker>
-    </tracker_ids>
-</project>';
-        $this->assertEquals($this->formatXml($xml), $this->formatXml($res['data']));
+        $this->assertEquals('POST', $response['method']);
+        $this->assertEquals('/projects.xml', $response['path']);
+        $this->assertXmlStringEqualsXmlString(
+            <<< XML
+            <?xml version="1.0"?>
+            <project>
+                <name>some name</name>
+                <identifier>the_identifier</identifier>
+                <tracker_ids type="array">
+                    <tracker>1</tracker>
+                    <tracker>2</tracker>
+                    <tracker>3</tracker>
+                </tracker_ids>
+            </project>
+            XML,
+            $response['data']
+        );
     }
 
     public function testUpdate()
     {
-        $api = $this->client->getApi('project');
+        /** @var \Redmine\Api\Project */
+        $api = MockClient::create()->getApi('project');
         $res = $api->update(1, [
             'name' => 'different name',
         ]);
-        $res = json_decode($res, true);
+        $response = json_decode($res, true);
 
-        $xml = '<?xml version="1.0"?>
-<project>
-    <id>1</id>
-    <name>different name</name>
-</project>';
-        $this->assertEquals($this->formatXml($xml), $this->formatXml($res['data']));
-    }
-
-    private function formatXml($xml)
-    {
-        $dom = new DOMDocument('1.0');
-        $dom->preserveWhiteSpace = false;
-        $dom->formatOutput = true;
-        $dom->loadXML((new SimpleXMLElement($xml))->asXML());
-
-        return $dom->saveXML();
+        $this->assertEquals('PUT', $response['method']);
+        $this->assertEquals('/projects/1.xml', $response['path']);
+        $this->assertXmlStringEqualsXmlString(
+            <<< XML
+            <?xml version="1.0"?>
+            <project>
+                <id>1</id>
+                <name>different name</name>
+            </project>
+            XML,
+            $response['data']
+        );
     }
 }