+++ tests/Integration/UserXmlTest.php
@@ -2,27 +2,16 @@
 
 namespace Redmine\Tests\Integration;
 
-use DOMDocument;
 use PHPUnit\Framework\TestCase;
 use Redmine\Exception\MissingParameterException;
 use Redmine\Tests\Fixtures\MockClient;
-use SimpleXMLElement;
 
 class UserXmlTest extends TestCase
 {
-    /**
-     * @var MockClient
-     */
-    private $client;
-
-    public function setup(): void
-    {
-        $this->client = new MockClient('http://test.local', 'asdf');
-    }
-
     public function testCreateBlank()
     {
-        $api = $this->client->getApi('user');
+        /** @var \Redmine\Api\User */
+        $api = MockClient::create()->getApi('user');
         $this->assertInstanceOf('Redmine\Api\User', $api);
 
         $this->expectException(MissingParameterException::class);
@@ -33,48 +22,52 @@ class UserXmlTest extends TestCase
 
     public function testCreateComplex()
     {
-        $api = $this->client->getApi('user');
+        /** @var \Redmine\Api\User */
+        $api = MockClient::create()->getApi('user');
         $res = $api->create([
             'login' => 'test',
             'firstname' => 'test',
             'lastname' => 'test',
             'mail' => 'test@example.com',
         ]);
-        $res = json_decode($res, true);
+        $response = json_decode($res, true);
 
-        $xml = '<?xml version="1.0"?>
-<user>
-    <login>test</login>
-    <lastname>test</lastname>
-    <firstname>test</firstname>
-    <mail>test@example.com</mail>
-</user>';
-        $this->assertEquals($this->formatXml($xml), $this->formatXml($res['data']));
+        $this->assertEquals('POST', $response['method']);
+        $this->assertEquals('/users.xml', $response['path']);
+        $this->assertXmlStringEqualsXmlString(
+            <<< XML
+            <?xml version="1.0"?>
+            <user>
+                <login>test</login>
+                <lastname>test</lastname>
+                <firstname>test</firstname>
+                <mail>test@example.com</mail>
+            </user>
+            XML,
+            $response['data']
+        );
     }
 
     public function testUpdate()
     {
-        $api = $this->client->getApi('user');
+        /** @var \Redmine\Api\User */
+        $api = MockClient::create()->getApi('user');
         $res = $api->update(1, [
             'firstname' => 'Raul',
         ]);
-        $res = json_decode($res, true);
+        $response = json_decode($res, true);
 
-        $xml = '<?xml version="1.0"?>
-<user>
-    <id>1</id>
-    <firstname>Raul</firstname>
-</user>';
-        $this->assertEquals($this->formatXml($xml), $this->formatXml($res['data']));
-    }
-
-    private function formatXml($xml)
-    {
-        $dom = new DOMDocument('1.0');
-        $dom->preserveWhiteSpace = false;
-        $dom->formatOutput = true;
-        $dom->loadXML((new SimpleXMLElement($xml))->asXML());
-
-        return $dom->saveXML();
+        $this->assertEquals('PUT', $response['method']);
+        $this->assertEquals('/users/1.xml', $response['path']);
+        $this->assertXmlStringEqualsXmlString(
+            <<< XML
+            <?xml version="1.0"?>
+            <user>
+                <id>1</id>
+                <firstname>Raul</firstname>
+            </user>
+            XML,
+            $response['data']
+        );
     }
 }